# docs/swagger.yaml
openapi: 3.0.3
info:
  title: GPTless Actual Intelligence API
  version: 1.0.0
servers:
  - url: https://api.gptless.au
  - url: https://preview.gptless.au
...
security:
  - AccessTokenAuth: []
paths:
...
  /v1/tokens:
    get:
      tags: [Tokens]
      summary: lists your token balance
      description: |
        Returns the current token balance for your account.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountName:
                  type: string
                  example: WamWarriors
              required:
                - accountName

      responses:
        "200":
          description: |
            **OK**

            Returns an object with the current token balance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountName:
                    type: string
                    example: Guy
                  tokenBalance:
                    type: number
                    example: 250

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to make these changes. These paths are developer only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            account cannot be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: accountName cannot be found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/tokens/buy:
    post:
      tags: [Tokens]
      summary: adds tokens to your account
      description: |
        Adds tokens to the given account and sends out an invoice using our invoicing API.

      parameters:
        - $ref: "#/components/parameters/AccessToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountName:
                  type: string
                  example: Guy
                tokens:
                  type: number
                  example: 100
              required:
                - accountName
                - tokens

      responses:
        "200":
          description: |
            **OK**

            Returns the updated token balance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountName:
                    type: string
                    example: WamWarriors
                  tokensAdded:
                    type: number
                    example: 100
                  tokenBalance:
                    type: number
                    example: 350

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to make these changes. These paths are developer only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            Group name cannot be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: accountName cannot be found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/tokens/redeem:
    post:
      tags: [Tokens]
      summary: Redeems your tokens for money
      description: |
        Redeems tokens from your account for money.

      parameters:
        - $ref: "#/components/parameters/AccessToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountName:
                  type: string
                  example: WamWarriors
                tokens:
                  type: number
                  example: 100
              required:
                - accountName
                - tokens

      responses:
        "200":
          description: |
            **OK**

            Returns the updated token balance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountName:
                    type: string
                    example: WamWarriors
                  tokensRedeemed:
                    type: number
                    example: 100
                  tokenBalance:
                    type: number
                    example: 150

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to make these changes. These paths are developer only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            Group name cannot be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: accountName cannot be found

        "409":
          description: |
            **Conflict**
            
            Not enough tokens to redeem.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                CONFLICT:
                  summary: CONFLICT
                  value:
                    error: CONFLICT
                    message: Not enough tokens to redeem

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
...
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: header
      name: AccessToken
  parameters:
...
    AccessToken:
      name: AccessToken
      in: header
      required: true
      schema:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
      description: AccessToken
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string